feat(tui): move layout_strategy to kanban-view, split view_strategy seam
Moves layout_strategy.rs (LayoutStrategy trait + SingleListLayout/
ColumnListsLayout/VirtualUnifiedLayout impls, ColumnBoundary,
build_query) verbatim into kanban_view::layout_strategy — it only
depended on ViewRefreshContext and CardList, both already available or
now co-located in kanban-view.

Splits view_strategy.rs along the render/no-render line already
present in UnifiedViewStrategy: ViewRefreshContext and the render-free
ViewStrategy trait (get_active_task_list, navigate_left/right,
refresh_task_lists, try_navigate_to_column, plus as_any/as_any_mut —
load-bearing for the Box<dyn ViewStrategy> downcasts in render_strategy.rs,
board_handlers.rs, navigation_handlers.rs and ui/main_view.rs) move to
kanban_view::view_strategy unchanged. UnifiedViewStrategy itself stays
in kanban-tui as a thin wrapper composing a Box<dyn LayoutStrategy>
(from kanban-view) with its own Box<dyn RenderStrategy> — its public
API (flat/grouped/kanban, get_layout_strategy, try_set_active_column_index)
is untouched.

Repoints every real consumer (grepped src/ and tests/ for
LayoutStrategy, ViewStrategy, ViewRefreshContext, ColumnBoundary,
SingleListLayout, ColumnListsLayout, VirtualUnifiedLayout, checking
for glob re-exports — none found for these two modules, they were
plain `pub mod` in lib.rs):
- render_strategy.rs, handlers/board_handlers.rs,
  handlers/navigation_handlers.rs: layout types now from
  kanban_view::layout_strategy; UnifiedViewStrategy stays local.
- app/view.rs, app/view_management.rs: ViewStrategy/ViewRefreshContext
  now from kanban_view::view_strategy; UnifiedViewStrategy stays local.
- lib.rs: drops the deleted `pub mod layout_strategy;`.

The KAN-1056 characterization test added ahead of this split
(test_unified_view_strategy_delegates_to_kanban_view_layout_strategy)
passes unchanged, proving the delegation wrapper altered no observable
behavior.
