+++ crates/kanban-tui/src/render_strategy.rs
@@ -3,9 +3,9 @@ use crate::components::{
     card_list_item::{render_card_list_item, CardListItemConfig},
     PanelConfig,
 };
-use crate::layout_strategy::ColumnBoundary;
 use crate::theme::{deleted_view_focused_border, label_text};
 use kanban_domain::card_lifecycle::sorted_board_columns;
+use kanban_view::layout_strategy::ColumnBoundary;
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
 use ratatui::text::{Line, Span};
 use ratatui::widgets::Paragraph;
@@ -68,8 +68,8 @@ impl RenderStrategy for SinglePanelRenderer {
 
                 if self.show_column_headers {
                     if let Some(task_list) = active_task_list {
-                        use crate::layout_strategy::VirtualUnifiedLayout;
                         use crate::view_strategy::UnifiedViewStrategy;
+                        use kanban_view::layout_strategy::VirtualUnifiedLayout;
 
                         // Try to get column boundaries from VirtualUnifiedLayout
                         let column_boundaries = app
