+++ crates/kanban-view/src/lib.rs
@@ -8,6 +8,7 @@ pub mod card_list;
 pub mod card_list_component;
 pub mod filter_state;
 pub mod filters;
+pub mod layout_strategy;
 pub mod list_component;
 pub mod list_nav;
 pub mod model;
@@ -15,5 +16,6 @@ pub mod scroll_indicators;
 pub mod search;
 pub mod selection_dialog;
 pub mod sprint_assign_list;
+pub mod view_strategy;
 
 pub use list_component::{ListComponent, ListRenderInfo};
